requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
azure-cognitiveservices-speech==1.34.0
elevenlabs==0.2.26
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
pydub==0.25.1